requirements.txt
aniso8601==1.2.0
appdirs==1.4.3
click==6.7
Flask==0.11.1
Flask-Cache==0.13.1
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.5
flask-restful-swagger-2==0.33
Flask-SQLAlchemy==2.1
flask-swagger==0.2.13
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==0.23
mysqlclient==1.3.9
packaging==16.8
PyJWT==1.4.1
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2016.10
PyYAML==3.12
redis==2.10.5
six==1.10.0
SQLAlchemy==1.1.5